Clarify the Codex adapter boundary
The cache fix now lives behind a small compatibility facade while the implementation is split into auth, protocol, transport, constants, and utility modules. This keeps the public import path stable and makes the WebSocket/HTTP continuity contract easier to audit separately from login and SSE parsing.

Constraint: External callers must continue importing from src/models/codex_chatgpt.js

Rejected: Leave the adapter as one monolithic file | the cache-lineage and WebSocket semantics were too easy to obscure among login and curl helpers

Confidence: high

Scope-risk: narrow

Directive: Keep src/models/codex_chatgpt.js as the unified public facade when adding Codex protocol internals

Tested: node --test tests/codex_chatgpt.test.js

Tested: npx eslint src/models/codex_chatgpt.js src/models/codex_chatgpt/*.js tests/codex_chatgpt.test.js

Tested: npm test
